success rate for time exercise no longer shows deciamals

+++ app/src/main/java/com/michaelcarrano/seven_min_workout/WorkoutCompleteActivity.java
@@ -107,7 +107,7 @@ public class WorkoutCompleteActivity extends AppCompatActivity {
                         if (te.isCompletedLastTime()) {
                             te.setTotalCompleted(te.getTotalCompleted() + 1);
                         }
-                        te.setCompletedPercentage(((double)te.getTotalCompleted() / (double)te.getWorkoutsCompleted()) * 100);
+                        te.setCompletedPercentage((int)(((double)te.getTotalCompleted() / (double)te.getWorkoutsCompleted()) * 100));
                     }
                 }
